+++ src/model.ts
@@ -12,6 +12,7 @@ export interface Scope {
   resolve<N extends Node>(qualifiedName: Name, allowLookup?: boolean): N | undefined
   include(...others: Scope[]): void
   register(...contributions: [Name, Node][]): void
+  localContributions(): [Name, Node][]
 }
 
 
@@ -94,7 +95,10 @@ export abstract class Node {
   }
 
   get categories(): Function[] { return [this.constructor] }
-  get sourceFileName(): string | undefined { return this.parent.sourceFileName }
+  get sourceFileName(): string | undefined {
+    const parent = getPotentiallyUninitializedLazy(this, 'parent')
+    return parent?.sourceFileName
+  }
   get sourceInfo(): string { return `${this.sourceFileName ?? '--'}:${this.sourceMap?.start.line ?? '--'}` }
   get label(): string { return `[${this.kind}]{${this.id?.slice(-6) ?? '--'}} at ${this.sourceInfo}` }