+++ test/game.test.ts
@@ -3,45 +3,47 @@ import { resolve } from 'path'
 import { buildEnvironment } from './assertions'
 import natives from '../src/wre/wre.natives'
 import { Environment } from '../src'
-import interpret from '../src/interpreter/interpreter'
+import interpret, { Interpreter } from '../src/interpreter/interpreter'
 
 should()
 
-// TODO: Move the wollok code to language
+// TODO: Move the wollok code to language -> We need to run programs!
 
 describe('Wollok Game', () => {
 
   describe('actions', () => {
 
     let environment: Environment
+    let interpreter: Interpreter 
+
 
     before(async () => {
       environment = await buildEnvironment('**/*.wpgm', resolve('language', 'test', 'game'))
     })
 
+    beforeEach(() => {
+      interpreter = interpret(environment, natives)
+    })
+
     it('addVisual', () => {
-      const interpreter = interpret(environment, natives)
       interpreter.run('actions.addVisual')
       const visuals = interpreter.object('wollok.game.game').get('visuals')!.innerValue!
       visuals.should.have.length(1)
     })
 
     it('removeVisual', () => {
-      const interpreter = interpret(environment, natives)
       interpreter.run('actions.removeVisual')
       const visuals = interpreter.object('wollok.game.game').get('visuals')!.innerValue!
       visuals.should.have.length(0)
     })
 
     it('say', () => {
-      const interpreter = interpret(environment, natives)
       interpreter.run('actions.say')
       interpreter.object('actions.visual').get('message')!.innerValue!.should.equal('Hi!')
       interpreter.object('actions.visual').get('messageTime')!.innerValue!.should.equal(2000)
     })
 
     it('clear', () => {
-      const interpreter = interpret(environment, natives)
       interpreter.run('actions.clear')
       const visuals = interpreter.object('wollok.game.game')!.get('visuals')!.innerValue!
       visuals.should.have.length(0)
@@ -49,10 +51,13 @@ describe('Wollok Game', () => {
     })
 
     it('flush event', () => {
-      const interpreter = interpret(environment, natives)
       const game = interpreter.object('wollok.game.game')!
       const time = interpreter.reify(1)
       interpreter.send('flushEvents', game, time)
+    })
+
+    it('with file name game (devil test)', () => {
+      interpreter.run('game.juego')
     })
   })
 })